+++ met-web/src/services/userService/index.ts
@@ -1,7 +1,7 @@
 import { _kc } from 'constants/tenantConstants';
 import { userToken, userDetails, userAuthorization, userAuthentication } from './userSlice';
 import { Action, AnyAction, Dispatch } from 'redux';
-import jwt from 'jsonwebtoken';
+import _jwt from 'jsonwebtoken';
 import { UserDetail } from './types';
 import { AppConfig } from 'config';
 import Endpoints from 'apiManager/endpoints';
@@ -22,6 +22,7 @@ const initKeycloak = async (dispatch: Dispatch<AnyAction>) => {
             pkceMethod: 'S256',
             checkLoginIframe: false,
         });
+
         if (!authenticated) {
             console.warn('not authenticated!');
             dispatch(userAuthentication(authenticated));
@@ -48,7 +49,8 @@ const initKeycloak = async (dispatch: Dispatch<AnyAction>) => {
     }
 };
 
-let refreshInterval: NodeJS.Timer;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+let refreshInterval: any;
 const refreshToken = (dispatch: Dispatch<Action>) => {
     refreshInterval = setInterval(async () => {
         if (KeycloakData) {
@@ -63,31 +65,6 @@ const refreshToken = (dispatch: Dispatch<Action>) => {
             }
         }
     }, 60000);
-};
-
-// eslint-disable-next-line
-const authenticateAnonymouslyOnFormio = () => {
-    const user = AppConfig.formio.anonymousUser;
-    const roles = [AppConfig.formio.anonymousId];
-    authenticateFormio(user, roles);
-};
-
-const authenticateFormio = async (user: string, roles: string[]) => {
-    const FORMIO_TOKEN = jwt.sign(
-        {
-            external: true,
-            form: {
-                _id: AppConfig.formio.userResourceFormId, // form.io form Id of user resource
-            },
-            user: {
-                _id: user, // keep it like that
-                roles: roles,
-            },
-        },
-        AppConfig.formio.jwtSecret,
-    ); // TODO Move JWT secret key to COME From ENV
-
-    localStorage.setItem('formioToken', FORMIO_TOKEN);
 };
 
 /**
